Cleaning up old code.
